Submitted file of Project-1 DO (Rocket Landing)
In accordance with trail-4, I was not able to get an approximate solution for the angular acceleration. Henceforth, made final changes in the code. Implemented the drag function properly and added some layers in the optimizer which gave a precise output. In accordance with trail-3, as the number of iterations and hidden layer matters while optimization, I have introduced a hidden layer and increased the number of iterations for the loss.